refactor: divide tag parser to util package
- 확장성 있는 구조를 위해 수정하였습니다.
