+++ api/document-read-api/src/main/java/com/jachwirus/documentreadapi/util/HashTagParser.java
@@ -1,4 +1,4 @@
-package com.jachwirus.documentreadapi.dto.mapper;
+package com.jachwirus.documentreadapi.util;
 
 import com.jachwirus.documentreadapi.model.DocumentHashTag;
 import com.jachwirus.documentreadapi.model.HashTag;
@@ -6,8 +6,8 @@ import com.jachwirus.documentreadapi.model.HashTag;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class Util {
-    public static List<String> getHashTags(List<DocumentHashTag> tags) {
+public class HashTagParser {
+    public static List<String> extractHashTagsName(List<DocumentHashTag> tags) {
         return tags.stream()
                 .map(DocumentHashTag::getHashTag)
                 .map(HashTag::getTagName)
